feat(api): move #findPaginatedFilteredOrganizations controller method to right context
