+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -90,115 +90,6 @@ describe('Acceptance | Application | organization-controller', function () {
     });
   });
 
-  describe('GET /api/admin/organizations', function () {
-    let server;
-    let options;
-
-    beforeEach(async function () {
-      server = await createServer();
-
-      const userSuperAdmin = databaseBuilder.factory.buildUser.withRole();
-
-      databaseBuilder.factory.buildOrganization({
-        name: 'The name of the organization',
-        type: 'SUP',
-        externalId: '1234567A',
-      });
-      databaseBuilder.factory.buildOrganization({
-        name: 'Organization of the night',
-        type: 'PRO',
-        externalId: '1234568A',
-      });
-
-      options = {
-        method: 'GET',
-        url: '/api/admin/organizations',
-        payload: {},
-        headers: generateAuthenticatedUserRequestHeaders({ userId: userSuperAdmin.id }),
-      };
-
-      return databaseBuilder.commit();
-    });
-
-    describe('Resource access management', function () {
-      it('should respond with a 401 - unauthorized access - if user is not authenticated', async function () {
-        // given
-        options.headers.authorization = 'invalid.access.token';
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(401);
-      });
-
-      it('should respond with a 403 - forbidden access - if user has not role Super Admin', async function () {
-        // given
-        const nonSuperAdminUserId = 9999;
-        options.headers = generateAuthenticatedUserRequestHeaders({ userId: nonSuperAdminUserId });
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(403);
-      });
-    });
-
-    describe('Success case', function () {
-      it('should return a 200 status code response with JSON API serialized', async function () {
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.data).to.have.lengthOf(2);
-        expect(response.result.data[0].type).to.equal('organizations');
-      });
-
-      it('should return pagination meta data', async function () {
-        // given
-        const expectedMetaData = { page: 1, pageSize: 10, rowCount: 2, pageCount: 1 };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-      });
-
-      it('should return a 200 status code with paginated and filtered data', async function () {
-        // given
-        options.url = '/api/admin/organizations?filter[name]=orga&filter[externalId]=A&page[number]=2&page[size]=1';
-        const expectedMetaData = { page: 2, pageSize: 1, rowCount: 2, pageCount: 2 };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-        expect(response.result.data).to.have.lengthOf(1);
-        expect(response.result.data[0].type).to.equal('organizations');
-      });
-
-      it('should return a 200 status code with empty result', async function () {
-        // given
-        options.url =
-          '/api/admin/organizations?filter[name]=orga&filter[type]=sco&filter[externalId]=B&page[number]=1&page[size]=1';
-        const expectedMetaData = { page: 1, pageSize: 1, rowCount: 0, pageCount: 0 };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.meta).to.deep.equal(expectedMetaData);
-        expect(response.result.data).to.have.lengthOf(0);
-      });
-    });
-  });
-
   describe('GET /api/organizations/{id}/campaigns', function () {
     let campaignsData;
     let organizationId, otherOrganizationId;
